requirements.txt
accelerate>=1.12.0
hf_xet>=1.2.0
huggingface-hub>=0.36.0
imagehash>=4.3.1
numpy==1.26.4
openpyxl>=3.1.3
pandas>=2.3.3
pillow>=11.3.0
pytesseract>=0.3.13
qwen_vl_utils>=0.0.13
requests>=2.32.5
torch>=2.9.0
torchvision>=0.24.0
# transformers>=4.57.0